+++ packages/design/jobberStyle.js
@@ -2,6 +2,8 @@
 const fs = require("fs");
 // eslint-disable-next-line import/no-internal-modules
 const customPropertiesObject = require("./src/foundation.js");
+// eslint-disable-next-line import/no-internal-modules
+const { getShadowStyles } = require("./src/getMobileShadows.js");
 
 const regexExpressions = {
   cssVars: /var\((.*)\)/,
@@ -10,6 +12,7 @@ const regexExpressions = {
   rgbaVars: /rgba\(var\((.*)\),?(.*)\)/,
   removeAllNonNumerals: /[^0-9.+\-/*]/gi,
   extractAllVarGroups: /var\(.*?\)/g,
+  extractTimeNumber: /(\d+)ms/,
 };
 
 const customProperties = customPropertiesObject.customProperties;
@@ -24,9 +27,7 @@ fs.writeFile("./foundation.js", jsonContent, "utf8", function (err) {
     console.log("An error occured while writing JSON object to File.");
     return console.log(err);
   }
-  console.log("JSON file has been saved.");
 });
-
 const scssColors = getResolvedSCSSVariables(resolvedCssVars);
 
 fs.writeFile(
@@ -38,9 +39,9 @@ fs.writeFile(
       console.log("An error occured while writing SCSS to File.");
       return console.log(err);
     }
-    console.log("SCSS file has been saved.");
   },
 );
+writeMobileFoundationFiles();
 
 /**
  * Recursively resolve css custom properties.
@@ -61,6 +62,8 @@ fs.writeFile(
  * ```
  */
 
+// Added up to 13 statements to accommodate sharing timing with mobile
+/*eslint max-statements: ["error", 13]*/
 function jobberStyle(styling) {
   const styleValue = removeNewLines(customProperties[styling]);
 
@@ -70,9 +73,11 @@ function jobberStyle(styling) {
   const rgbVarRegexResult = regexExpressions.rgbVars.exec(styleValue);
   //rgbaVarRegexResult returns --base-unit and alpha (if exists) from rgba(var(--base-unit), alpha)
   const rgbaVarRegexResult = regexExpressions.rgbaVars.exec(styleValue);
-
   //calcRegexResult returns var(--base-unit) / 16 from calc(var(--base-unit) / 16)
   const calcRegexResult = regexExpressions.calculations.exec(styleValue);
+  //timeNumberResult returns 100 from "100ms"
+  const timeNumberResult = regexExpressions.extractTimeNumber.exec(styleValue);
+
   if (calcRegexResult) {
     return handleCalc(calcRegexResult);
   } else if (rgbVarRegexResult) {
@@ -81,6 +86,8 @@ function jobberStyle(styling) {
     return handleRbga(rgbaVarRegexResult);
   } else if (varRegexResult) {
     return jobberStyle(varRegexResult[1]);
+  } else if (timeNumberResult) {
+    return handleTiming(timeNumberResult[1]);
   } else {
     return isSpacingValue(styleValue) || isFloatValue(styleValue)
       ? parseFloat(styleValue)
@@ -102,6 +109,10 @@ function handleCalc(calcRegexResult) {
   return isSpacingValue(calculatedValue)
     ? parseFloat(calculatedValue)
     : calculatedValue;
+}
+
+function handleTiming(timeNumberResult) {
+  return parseFloat(timeNumberResult);
 }
 
 function handleRbga(rgbaVarRegexResult) {
@@ -247,4 +258,63 @@ function getPropertyValue(cssVar) {
 function removeNewLines(text) {
   if (!text) return text;
   return text.replace(/(\r\n|\n|\r)/gm, "");
+}
+
+function writeMobileFoundationFiles() {
+  const { androidShadows, iOSShadows } = getShadowStyles(resolvedCssVars);
+  const androidFoundationJobberStyle = {
+    ...resolvedCssVars,
+    ...androidShadows,
+  };
+  const iOSFoundationJobberStyle = {
+    ...resolvedCssVars,
+    ...iOSShadows,
+  };
+  const androidFoundationsExportString = `export const JobberStyle = ${JSON.stringify(
+    androidFoundationJobberStyle,
+    undefined,
+    2,
+  )}`;
+
+  const iOSFoundationsExportString = `export const JobberStyle = ${JSON.stringify(
+    iOSFoundationJobberStyle,
+    undefined,
+    2,
+  )}`;
+  fs.writeFile(
+    "./foundation.android.js",
+    androidFoundationsExportString,
+    "utf-8",
+    err => {
+      if (err) {
+        console.log("An error occured while writing Android Foundation File.");
+        return console.log(err);
+      }
+      console.log("Wrote Android Foundations file");
+    },
+  );
+  fs.writeFile(
+    "./foundation.ios.js",
+    iOSFoundationsExportString,
+    "utf-8",
+    err => {
+      if (err) {
+        console.log("An error occured while writing iOS Foundation File.");
+        return console.log(err);
+      }
+      console.log("Wrote iOS Foundations file");
+    },
+  );
+  fs.writeFile(
+    "./foundation.native.js",
+    iOSFoundationsExportString,
+    "utf-8",
+    err => {
+      if (err) {
+        console.log("An error occured while writing iOS Foundation File.");
+        return console.log(err);
+      }
+      console.log("Wrote Native Foundations file");
+    },
+  );
 }